Extracts Job object to controller and model
