+++ src/models/Job.js
@@ -1,0 +1,13 @@
+let data = [];
+
+module.exports = {
+    get() {
+        return data;
+    },
+    update(newJobs) {
+        data = newJobs;
+    },
+    delete(id) {
+        data = data.filter(job => Number(job.id) !== Number(id));
+    }
+};
